+++ Caribou/Components/SpecifyFeaturesComponent.cs
@@ -1,13 +1,10 @@
 ﻿namespace Caribou.Components
 {
     using System;
-    using System.Collections.Generic;
     using Grasshopper.Kernel;
+    using Caribou.Models;
     using Caribou.Properties;
     using Caribou.Forms;
-    using Caribou.Models;
-    using Caribou.Forms.Models;
-    using Eto.Forms;
 
     /// <summary>Provides a GUI interface to selecting/specifying predefined OSM features/subfeatures.</summary>
     public class SpecifyFeaturesComponent : BasePickerComponent
@@ -15,37 +12,10 @@
 
         public SpecifyFeaturesComponent() : base("Specify Features", "OSM Specify",
             "Provides a graphical interface to specify a list of OSM features that the Extract components will then find.", "Select")
-        {
-            // Setup form-items for tags provided and parsed into OSM/Form objects
-            this.selectableOSMs = new TreeGridItemCollection();
-            var indexOfParents = new Dictionary<string, int>();
-
-            var primaryFeatures = new List<OSMMetaData>(OSMDefinedFeatures.Primary.Values);
-            for (var i = 0; i < primaryFeatures.Count; i++)
-            {
-                var parentItem = new CaribouTreeGridItem(primaryFeatures[i], 0, 0, false);
-
-                // Insert untagged item
-                var description = $"Items that are specified as {primaryFeatures[i].Name}, but without more specific subfeature information";
-                var childUntaggedOSM = new OSMMetaData("yes", "", 
-                                                       description, primaryFeatures[i]);
-                var childUntaggedItem = new CaribouTreeGridItem(childUntaggedOSM, 0, 0, false);
-                parentItem.Children.Add(childUntaggedItem);
-
-                this.selectableOSMs.Add(parentItem);
-                indexOfParents[primaryFeatures[i].TagType] = i;
-            }
-
-            foreach (var item in OSMDefinedFeatures.SubFeatures())
-            {
-                var parentItem = this.selectableOSMs[indexOfParents[item["feature"]]] as CaribouTreeGridItem;
-
-                var childOSM = new OSMMetaData(item["subfeature"], null, item["description"], 
-                    parentItem.OSMData);
-                var childItem = new CaribouTreeGridItem(childOSM, int.Parse(item["nodes"]), int.Parse(item["ways"]), false);
-                parentItem.Children.Add(childItem);
-            }
+        {            
+            this.selectableOSMs = OSMDefinedFeatures.GetTreeCollection(); // Setup form-items for tags provided and parsed into OSM/Form objects
         }
+
 
         #region InOut Params
         protected override void RegisterInputParams(GH_InputParamManager pManager) { }